+++ src/main/kotlin/com/nextstep/racingcar/view/ResultView.kt
@@ -1,4 +1,4 @@
-package com.nextstep.racingcar
+package com.nextstep.racingcar.view
 
 class ResultView {
